refactor(landing): :sparkles: renaming icon
