+++ src/components/icons/partners/UniversityLoyolaIcon.tsx
